+++ src/libiberty/objalloc.c
@@ -187,32 +187,32 @@ objalloc_free(struct objalloc *o)
    recently allocated blocks.  */
 
 void
-objalloc_free_block (struct objalloc *o, PTR block)
+objalloc_free_block(struct objalloc *o, PTR block)
 {
   struct objalloc_chunk *p, *small;
-  char *b = (char *) block;
+  char *b = (char *)block;
 
   /* First set P to the chunk which contains the block we are freeing,
      and set Q to the last small object chunk we see before P.  */
   small = NULL;
-  for (p = (struct objalloc_chunk *) o->chunks; p != NULL; p = p->next)
+  for (p = (struct objalloc_chunk *)o->chunks; p != NULL; p = p->next)
     {
       if (p->current_ptr == NULL)
 	{
-	  if (b > (char *) p && b < (char *) p + CHUNK_SIZE)
+	  if (b > (char *)p && b < (char *)p + CHUNK_SIZE)
 	    break;
 	  small = p;
 	}
       else
 	{
-	  if (b == (char *) p + CHUNK_HEADER_SIZE)
+	  if (b == (char *)p + CHUNK_HEADER_SIZE)
 	    break;
 	}
     }
 
   /* If we can't find the chunk, the caller has made a mistake.  */
   if (p == NULL)
-    abort ();
+    abort();
 
   if (p->current_ptr == NULL)
     {
@@ -227,7 +227,7 @@ objalloc_free_block (struct objalloc *o, PTR block)
 	 can then reset the new current_ptr to B.  */
 
       first = NULL;
-      q = (struct objalloc_chunk *) o->chunks;
+      q = (struct objalloc_chunk *)o->chunks;
       while (q != p)
 	{
 	  struct objalloc_chunk *next;
@@ -237,10 +237,10 @@ objalloc_free_block (struct objalloc *o, PTR block)
 	    {
 	      if (small == q)
 		small = NULL;
-	      free (q);
+	      free(q);
 	    }
 	  else if (q->current_ptr > b)
-	    free (q);
+	    free(q);
 	  else if (first == NULL)
 	    first = q;
 
@@ -249,11 +249,11 @@ objalloc_free_block (struct objalloc *o, PTR block)
 
       if (first == NULL)
 	first = p;
-      o->chunks = (PTR) first;
+      o->chunks = (PTR)first;
 
-      /* Now start allocating from this small block again.  */
+      /* Now start allocating from this small block again: */
       o->current_ptr = b;
-      o->current_space = ((char *) p + CHUNK_SIZE) - b;
+      o->current_space = (unsigned long)(((char *)p + CHUNK_SIZE) - b);
     }
   else
     {
@@ -269,23 +269,24 @@ objalloc_free_block (struct objalloc *o, PTR block)
       current_ptr = p->current_ptr;
       p = p->next;
 
-      q = (struct objalloc_chunk *) o->chunks;
+      q = (struct objalloc_chunk *)o->chunks;
       while (q != p)
 	{
 	  struct objalloc_chunk *next;
 
 	  next = q->next;
-	  free (q);
+	  free(q);
 	  q = next;
 	}
 
-      o->chunks = (PTR) p;
+      o->chunks = (PTR)p;
 
       while (p->current_ptr != NULL)
 	p = p->next;
 
       o->current_ptr = current_ptr;
-      o->current_space = (((char *)p + CHUNK_SIZE) - current_ptr);
+      o->current_space = (unsigned long)(((char *)p + CHUNK_SIZE)
+                                         - current_ptr);
     }
 }
