+++ src/libiberty/vsnprintf.c
@@ -77,10 +77,10 @@ vsnprintf(char *s, size_t n, const char *format, va_list ap)
   if (n > 0)
     {
       if ((long)n > result)
-	memcpy(s, buf, (result + 1));
+	memcpy(s, buf, ((size_t)result + 1UL));
       else
         {
-	  memcpy(s, buf, (n - 1));
+	  memcpy(s, buf, (n - 1UL));
 	  s[n - 1] = 0;
 	}
     }
